# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Project:
    properties:
      api_style:
        type: string
      created_at:
        type: string
      database_type:
        type: string
      description:
        type: string
      id:
        type: string
      language:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  domain.Schema:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      project_id:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  handlers.CreateProjectRequest:
    properties:
      api_style:
        type: string
      database_type:
        type: string
      description:
        type: string
      language:
        type: string
      name:
        type: string
    required:
    - api_style
    - database_type
    - language
    - name
    type: object
  handlers.GenerateProjectRequest:
    properties:
      output_dir:
        type: string
    required:
    - output_dir
    type: object
host: localhost:7070
info:
  contact: {}
  description: Backend as a Service (BaaS) Platform API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Aurora API
  version: "1.0"
paths:
  /health:
    get:
      description: Get the health status and uptime of the Aurora service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Check service health
      tags:
      - health
  /projects:
    get:
      description: Get a list of all Backend as a Service projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/domain.Project'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new Backend as a Service project with specified configuration
      parameters:
      - description: Project creation request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Remove a project and all its associated schemas
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete project by ID
      tags:
      - projects
    get:
      description: Retrieve detailed information about a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update the configuration of an existing project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project update request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update project by ID
      tags:
      - projects
  /projects/{id}/generate:
    post:
      consumes:
      - application/json
      description: Trigger code generation for a specific project based on its latest
        schema
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Code generation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GenerateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate backend code
      tags:
      - codegen
  /projects/{id}/schemas:
    get:
      description: Get a list of all schema versions associated with a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/domain.Schema'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all schema versions
      tags:
      - schemas
    post:
      consumes:
      - text/plain
      description: Apply a new YAML schema definition to a project. This will increment
        the schema version.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: YAML schema content
        in: body
        name: schema
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Schema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Apply a new schema version
      tags:
      - schemas
  /projects/{id}/schemas/latest:
    get:
      description: Retrieve the most recent schema version for a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Schema'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get latest schema version
      tags:
      - schemas
swagger: "2.0"